refactor: use shared implementation of GetFlagValue
